+++ EppoEngine/Source/Platform/Vulkan/VulkanPipeline.h
@@ -1,0 +1,33 @@
+#pragma once
+
+#include "Platform/Vulkan/Vulkan.h"
+#include "Renderer/Buffer/UniformBufferSet.h"
+#include "Renderer/Pipeline.h"
+
+namespace Eppo
+{
+	class VulkanPipeline : Pipeline
+	{
+	public:
+		VulkanPipeline(const PipelineSpecification& specification);
+		~VulkanPipeline();
+
+		void UpdateDescriptors(uint32_t frameIndex, Ref<UniformBufferSet> uniformBufferSet);
+
+		VkPipeline GetPipeline() const { return m_Pipeline; }
+		VkPipelineLayout GetPipelineLayout() const { return m_PipelineLayout; }
+
+		const std::vector<VkDescriptorSet>& GetDescriptorSets(uint32_t frameIndex) { return m_DescriptorSets.at(frameIndex); }
+
+		const PipelineSpecification& GetSpecification() const override { return m_Specification; };
+
+	private:
+		PipelineSpecification m_Specification;
+
+		VkPipeline m_Pipeline;
+		VkPipelineLayout m_PipelineLayout;
+
+		std::vector<VkDescriptorPool> m_DescriptorPools;
+		std::unordered_map<uint32_t, std::vector<VkDescriptorSet>> m_DescriptorSets; // frame > set
+	};
+}
