+++ EppoEngine/Source/Platform/Vulkan/VulkanRenderer.h
@@ -2,13 +2,13 @@
 
 #include "Platform/Vulkan/Descriptor/DescriptorAllocator.h"
 #include "Platform/Vulkan/Descriptor/DescriptorLayoutCache.h"
-#include "Renderer/Renderer.h"
-#include "Renderer/ShaderLibrary.h"
+#include "Renderer/RendererAPI.h"
+#include "Renderer/Buffer/UniformBuffer.h"
 
 namespace Eppo
 {
 	// TODO: Do we still need all the const references with our new ref counter?
-	class VulkanRenderer : public Renderer
+	class VulkanRenderer : public RendererAPI
 	{
 	public:
 		VulkanRenderer() = default;
@@ -32,21 +32,11 @@ namespace Eppo
 		Ref<Shader> GetShader(const std::string& name) override;
 
 		// Geometry
-		void RenderGeometry(Ref<RenderCommandBuffer> renderCommandBuffer, Ref<Pipeline> pipeline, Ref<UniformBuffer> environmentUB, Ref<UniformBuffer> cameraUB, Ref<Mesh> mesh, const glm::mat4& transform) override;
+		void RenderGeometry(Ref<RenderCommandBuffer> renderCommandBuffer, Ref<Pipeline> pipeline, Ref<Mesh> mesh, const glm::mat4& transform) override;
 
 		// Descriptor sets
-		Ref<DescriptorAllocator> GetDescriptorAllocator() const { return m_DescriptorAllocator; }
-		Ref<DescriptorLayoutCache> GetDescriptorLayoutCache() const { return m_DescriptorLayoutCache; }
-		VkDescriptorSet AllocateDescriptorSet(VkDescriptorSetAllocateInfo& allocInfo) const;
-
-	private:
-		Scope<RenderCommandQueue> m_CommandQueue;
-		Ref<RenderCommandBuffer> m_CommandBuffer;
-
-		Ref<DescriptorAllocator> m_DescriptorAllocator;
-		Ref<DescriptorLayoutCache> m_DescriptorLayoutCache;
-		std::vector<VkDescriptorPool> m_DescriptorPools;
-
-		Scope<ShaderLibrary> m_ShaderLibrary;
+		static Ref<DescriptorAllocator> GetDescriptorAllocator();
+		static Ref<DescriptorLayoutCache> GetDescriptorLayoutCache();
+		static VkDescriptorSet AllocateDescriptorSet(VkDescriptorSetAllocateInfo& allocInfo) const;
 	};
 }
