+++ EppoEngine/Source/Renderer/Material.h
@@ -5,22 +5,13 @@
 
 namespace Eppo
 {
-	class Material
+	class Material : public RefCounter
 	{
 	public:
-		Material(Ref<Shader> shader);
-		~Material();
+		virtual ~Material() {}
 
-		void Set(const std::string& name, Ref<Texture> texture, uint32_t arrayIndex = 0);
-		VkDescriptorSet GetDescriptorSet(uint32_t imageIndex);
-		VkDescriptorSet GetCurrentDescriptorSet();
+		virtual void Set(const std::string& name, Ref<Texture> texture, uint32_t arrayIndex = 0) = 0;
 
-	private:
-		Ref<Shader> m_Shader;
-		Ref<Texture> m_Texture; // TODO: Can be multiple in case of meshes?
-
-		uint32_t m_Binding;
-
-		std::vector<VkDescriptorSet> m_DescriptorSets;
+		static Ref<Material> Create(Ref<Shader> shader);
 	};
 }
