+++ EppoEngine/Source/Renderer/Pipeline.h
@@ -4,6 +4,8 @@
 #include "Renderer/Shader.h"
 #include "Renderer/VertexBufferLayout.h"
 
+struct VkPushConstantRange;
+
 namespace Eppo
 {
 	struct PipelineSpecification
@@ -18,26 +20,13 @@ namespace Eppo
 		bool DepthTesting = false;
 	};
 
-	class Pipeline
+	class Pipeline : public RefCounter
 	{
 	public:
-		Pipeline(const PipelineSpecification& specification);
-		~Pipeline();
+		virtual ~Pipeline() {};
 
-		const PipelineSpecification& GetSpecification() const { return m_Specification; }
-
-		const std::vector<VkDescriptorSet>& GetDescriptorSets(uint32_t frameIndex) { return m_DescriptorSets.at(frameIndex); }
-
-		VkPipeline GetPipeline() const { return m_Pipeline; }
-		VkPipelineLayout GetPipelineLayout() const { return m_PipelineLayout; }
-
-	private:
-		PipelineSpecification m_Specification;
-
-		std::vector<VkDescriptorPool> m_DescriptorPools;
-		std::unordered_map<uint32_t, std::vector<VkDescriptorSet>> m_DescriptorSets; // frame > set
-
-		VkPipeline m_Pipeline;
-		VkPipelineLayout m_PipelineLayout;
+		virtual const PipelineSpecification& GetSpecification() const = 0;
+		
+		static Ref<Pipeline> Create(const PipelineSpecification& specification);
 	};
 }
