+++ EppoEngine/Source/Renderer/Renderer.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Renderer/Buffer/UniformBuffer.h"
+#include "Renderer/Buffer/UniformBufferSet.h"
 #include "Renderer/Mesh/Mesh.h"
 #include "Renderer/Pipeline.h"
 #include "Renderer/RenderCommandBuffer.h"
@@ -12,29 +12,24 @@ namespace Eppo
 	class Renderer
 	{
 	public:
-		virtual ~Renderer() {}
-
-		virtual void Init() = 0;
-		virtual void Shutdown() = 0;
+		static void Init();
+		static void Shutdown();
 
 		// Frame index
-		virtual uint32_t GetCurrentFrameIndex() const = 0;
+		static uint32_t GetCurrentFrameIndex();
 
 		// Render pass
-		virtual void BeginRenderPass(Ref<RenderCommandBuffer> renderCommandBuffer, Ref<Pipeline> pipeline) = 0;
-		virtual void EndRenderPass(Ref<RenderCommandBuffer> renderCommandBuffer) = 0;
+		static void BeginRenderPass(Ref<RenderCommandBuffer> renderCommandBuffer, Ref<Pipeline> pipeline);
+		static void EndRenderPass(Ref<RenderCommandBuffer> renderCommandBuffer);
 
 		// Render commands
-		virtual void ExecuteRenderCommands() = 0;
-		virtual void SubmitCommand(RenderCommand command) = 0;
+		static void ExecuteRenderCommands();
+		static void SubmitCommand(RenderCommand command);
 
 		// Shaders
-		virtual Ref<Shader> GetShader(const std::string& name) = 0;
+		static Ref<Shader> GetShader(const std::string& name);
 
 		// Geometry
-		virtual void RenderGeometry(Ref<RenderCommandBuffer> renderCommandBuffer, Ref<Pipeline> pipeline, Ref<UniformBuffer> environmentUB, Ref<UniformBuffer> cameraUB, Ref<Mesh> mesh, const glm::mat4& transform) = 0;
-	
-		// Create me
-		static Ref<Renderer> Create();
+		static void RenderGeometry(Ref<RenderCommandBuffer> renderCommandBuffer, Ref<Pipeline> pipeline, Ref<UniformBufferSet> uniformBufferSet, Ref<Mesh> mesh, const glm::mat4& transform);
 	};
 }
