+++ EppoEngine/Source/Renderer/RendererContext.h
@@ -4,7 +4,7 @@ struct GLFWwindow;
 
 namespace Eppo
 {
-	enum class RendererAPI
+	enum class RendererAPIType
 	{
 		OpenGL,
 		Vulkan
@@ -18,11 +18,11 @@ namespace Eppo
 		virtual void Init() = 0;
 		virtual void Shutdown() = 0;
 
-		static RendererAPI GetAPI() { return s_API; }
+		static RendererAPIType GetAPI() { return s_API; }
 		static Ref<RendererContext> Get();
 		static Ref<RendererContext> Create(void* windowHandle);
 
 	private:
-		static RendererAPI s_API;
+		static RendererAPIType s_API;
 	};
 }
