+++ EppoEngine/Source/Renderer/SceneRenderer.cpp
@@ -1,6 +1,8 @@
 #include "pch.h"
 #include "SceneRenderer.h"
 
+#include "Platform/Vulkan/VulkanContext.h" // TODO: This should absolutely not be here!
+#include "Platform/Vulkan/VulkanPipeline.h" // TODO: This should absolutely not be here!
 #include "Renderer/Renderer.h"
 #include "Scene/Scene.h"
 
@@ -12,25 +14,19 @@ namespace Eppo
 	SceneRenderer::SceneRenderer(Ref<Scene> scene, const RenderSpecification& renderSpecification)
 		: m_RenderSpecification(renderSpecification)
 	{
-		Ref<RendererContext> context = RendererContext::Get();
-		Ref<VulkanSwapchain> swapchain = context->GetSwapchain();
-		VkDevice device = context->GetLogicalDevice()->GetNativeDevice();
-
-		m_CommandBuffer = CreateRef<RenderCommandBuffer>();
+		m_CommandBuffer = Ref<RenderCommandBuffer>::Create();
 
 		// Geometry
 		{
 			FramebufferSpecification framebufferSpec;
 			framebufferSpec.Attachments = { ImageFormat::RGBA8, ImageFormat::Depth };
-			framebufferSpec.Width = swapchain->GetWidth();
-			framebufferSpec.Height = swapchain->GetHeight();
 			framebufferSpec.ClearColorOnLoad = true;
 			framebufferSpec.ClearColor = { 0.0f, 0.0f, 0.0f, 1.0f };
 			framebufferSpec.ClearDepthOnLoad = true;
 			framebufferSpec.ClearDepth = 1.0f;
 
 			PipelineSpecification pipelineSpec;
-			pipelineSpec.Framebuffer = CreateRef<Framebuffer>(framebufferSpec);
+			pipelineSpec.Framebuffer = Framebuffer::Create(framebufferSpec);
 			pipelineSpec.Shader = Renderer::GetShader("geometry");
 			pipelineSpec.Layout = {
 				{ ShaderDataType::Float3, "inPosition" },
@@ -43,7 +39,7 @@ namespace Eppo
 				{ VK_SHADER_STAGE_FRAGMENT_BIT, sizeof(glm::mat4), sizeof(glm::vec3) + sizeof(float) }
 			};
 
-			m_GeometryPipeline = CreateRef<Pipeline>(pipelineSpec);
+			m_GeometryPipeline = Pipeline::Create(pipelineSpec);
 		}
 
 		// Lighting
@@ -58,14 +54,12 @@ namespace Eppo
 
 			FramebufferSpecification framebufferSpec;
 			framebufferSpec.Attachments = { ImageFormat::Depth };
-			framebufferSpec.Width = swapchain->GetWidth();
-			framebufferSpec.Height = swapchain->GetHeight();
 			framebufferSpec.ClearColorOnLoad = false;
 			framebufferSpec.ClearDepthOnLoad = true;
 			framebufferSpec.ClearDepth = 1.0f;
 
 			PipelineSpecification pipelineSpec;
-			pipelineSpec.Framebuffer = CreateRef<Framebuffer>(framebufferSpec);
+			pipelineSpec.Framebuffer = Framebuffer::Create(framebufferSpec);
 			pipelineSpec.Shader = Renderer::GetShader("shadow");
 			pipelineSpec.Layout = {
 				{ ShaderDataType::Float3, "inPosition" },
@@ -77,21 +71,20 @@ namespace Eppo
 				{ VK_SHADER_STAGE_VERTEX_BIT, 0, sizeof(glm::mat4) }
 			};
 
-			m_ShadowPipeline = CreateRef<Pipeline>(pipelineSpec);
+			m_ShadowPipeline = Pipeline::Create(pipelineSpec);
 		}
 
 		// Uniform buffers
-		m_UniformBufferSet = CreateRef<UniformBufferSet>();
+		m_UniformBufferSet = Ref<UniformBufferSet>::Create();
 		// TODO: Set doesn't solve the problem of having one uniform buffer for multiple different pipeline layouts
-		/*for (uint32_t i = 0; i < VulkanConfig::MaxFramesInFlight; i++)
+		for (uint32_t i = 0; i < VulkanConfig::MaxFramesInFlight; i++)
 		{
-			m_UniformBufferSet->Set(CreateRef<UniformBuffer>(sizeof(CameraData), 0), i, 0);
-			m_UniformBufferSet->Set(CreateRef<UniformBuffer>(sizeof(EnvironmentData), 1), i, 0);
-			m_UniformBufferSet->Set(CreateRef<UniformBuffer>(sizeof(EnvironmentData), 2), i, 0);
-		}*/
+			m_UniformBufferSet->Set(UniformBuffer::Create(sizeof(CameraData), 0), i, 0);
+			m_UniformBufferSet->Set(UniformBuffer::Create(sizeof(EnvironmentData), 1), i, 0);
+		}
 		
-		m_CameraUniformBuffer = CreateRef<UniformBuffer>(sizeof(CameraData), 0);
-		m_EnvironmentUniformBuffer = CreateRef<UniformBuffer>(sizeof(EnvironmentData), 1);
+		m_CameraUniformBuffer = UniformBuffer::Create(sizeof(CameraData), 0);
+		m_EnvironmentUniformBuffer = UniformBuffer::Create(sizeof(EnvironmentData), 1);
 	}
 
 	void SceneRenderer::RenderGui()
@@ -190,53 +183,19 @@ namespace Eppo
 		uint32_t frameIndex = Renderer::GetCurrentFrameIndex();
 
 		// Descriptor sets
-		const auto& descriptorSets = m_GeometryPipeline->GetDescriptorSets(frameIndex);
-
-		std::vector<VkWriteDescriptorSet> writeDescriptors;
-		{
-			VkWriteDescriptorSet& writeDescriptor = writeDescriptors.emplace_back();
-			writeDescriptor.sType = VK_STRUCTURE_TYPE_WRITE_DESCRIPTOR_SET;
-			writeDescriptor.descriptorCount = 1;
-			writeDescriptor.descriptorType = VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER;
-			writeDescriptor.dstSet = descriptorSets[0];
-			writeDescriptor.dstBinding = 0;
-			writeDescriptor.pBufferInfo = &m_CameraUniformBuffer->GetDescriptorBufferInfo();
-		}
-
-		{
-			VkWriteDescriptorSet& writeDescriptor = writeDescriptors.emplace_back();
-			writeDescriptor.sType = VK_STRUCTURE_TYPE_WRITE_DESCRIPTOR_SET;
-			writeDescriptor.descriptorCount = 1;
-			writeDescriptor.descriptorType = VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER;
-			writeDescriptor.dstSet = descriptorSets[0];
-			writeDescriptor.dstBinding = 1;
-			writeDescriptor.pBufferInfo = &m_EnvironmentUniformBuffer->GetDescriptorBufferInfo();
-		}
-
-		{
-			Ref<Image> depthImage = m_ShadowPipeline->GetSpecification().Framebuffer->GetDepthImage();
-
-			VkWriteDescriptorSet& writeDescriptor = writeDescriptors.emplace_back();
-			writeDescriptor.sType = VK_STRUCTURE_TYPE_WRITE_DESCRIPTOR_SET;
-			writeDescriptor.descriptorCount = 1;
-			writeDescriptor.descriptorType = VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER;
-			writeDescriptor.dstSet = descriptorSets[0];
-			writeDescriptor.dstBinding = 2;
-			writeDescriptor.pImageInfo = &depthImage->GetDescriptorImageInfo();
-		}
-
-		VkDevice device = RendererContext::Get()->GetLogicalDevice()->GetNativeDevice();
-		vkUpdateDescriptorSets(device, writeDescriptors.size(), writeDescriptors.data(), 0, nullptr);
+		// TODO: Update materials through method
+		Ref<VulkanPipeline> vulkanPipeline = m_GeometryPipeline.As<VulkanPipeline>();
+		vulkanPipeline->UpdateDescriptors(frameIndex, m_UniformBufferSet);
 
 		EntityHandle handle;
 		for (auto& [entity, dc] : m_DrawList)
 		{
-			Renderer::RenderGeometry(m_CommandBuffer, m_GeometryPipeline, m_EnvironmentUniformBuffer, m_CameraUniformBuffer, dc.Mesh, dc.Transform);
+			Renderer::RenderGeometry(m_CommandBuffer, m_GeometryPipeline, m_UniformBufferSet, dc.Mesh, dc.Transform);
 			handle = entity;
 		}
 
 		glm::mat4 transform = glm::translate(glm::mat4(1.0f), glm::vec3(m_EnvironmentBuffer.LightPosition));
-		Renderer::RenderGeometry(m_CommandBuffer, m_GeometryPipeline, m_EnvironmentUniformBuffer, m_CameraUniformBuffer, m_DrawList[handle].Mesh, transform);
+		Renderer::RenderGeometry(m_CommandBuffer, m_GeometryPipeline, m_UniformBufferSet, m_DrawList[handle].Mesh, transform);
 
 		Renderer::EndRenderPass(m_CommandBuffer);
